# Improve average performance of LongTaskTimer for out-of-order stopping